+++ gui/CMakeLists.txt
@@ -8,54 +8,46 @@ find_package(boost_plugin_loader REQUIRED)
 
 # Add widget plugin
 add_library(
-  ${PROJECT_NAME}_widget SHARED
+  ${PROJECT_NAME}_gui SHARED
   # Utilities
-  include/widget/configurable_widget.h
-  src/widgets/configurable_widget.cpp
-  src/widgets/camera_intrinsics.ui
-  src/widgets/aspect_ratio_pixmap_label.cpp
-  include/widget/aspect_ratio_pixmap_label.h
+  src/configurable_widget.cpp
+  src/aspect_ratio_pixmap_label.cpp
+  include/industrial_calibration/gui/aspect_ratio_pixmap_label.h
   # Calibration widgets
-  src/widgets/camera_intrinsics.cpp
-  include/widget/camera_intrinsics.h
-  src/widgets/transform_guess.ui
-  src/widgets/transform_guess.cpp
-  include/widget/transform_guess.h
+  src/camera_intrinsics.ui
+  src/camera_intrinsics.cpp
+  src/transform_guess.ui
+  src/transform_guess.cpp
   # Target finders
-  src/widgets/charuco_grid_target_finder.ui
-  src/widgets/charuco_grid_target_finder.cpp
-  include/widget/charuco_grid_target_finder.h
-  src/widgets/aruco_grid_target_finder.ui
-  src/widgets/aruco_grid_target_finder.cpp
-  include/widget/aruco_grid_target_finder.h
-  src/widgets/modified_circle_grid_target_finder.ui
-  src/widgets/modified_circle_grid_target_finder.cpp
-  include/widget/modified_circle_grid_target_finder.h
+  src/charuco_grid_target_finder.ui
+  src/charuco_grid_target_finder.cpp
+  src/aruco_grid_target_finder.ui
+  src/aruco_grid_target_finder.cpp
+  src/modified_circle_grid_target_finder.ui
+  src/modified_circle_grid_target_finder.cpp
   # Extrinsic hand-eye calibration widget
-  src/widgets/extrinsic_hand_eye_calibration_widget.ui
-  src/widgets/extrinsic_hand_eye_calibration_widget.cpp
-  include/widget/extrinsic_hand_eye_calibration_widget.h
+  src/extrinsic_hand_eye_calibration_widget.ui
+  src/extrinsic_hand_eye_calibration_widget.cpp
 )
-target_include_directories(${PROJECT_NAME}_widget PUBLIC
+target_include_directories(${PROJECT_NAME}_gui PUBLIC
   "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>"
   "$<INSTALL_INTERFACE:include>")
-target_link_libraries(${PROJECT_NAME}_widget PUBLIC
+target_link_libraries(${PROJECT_NAME}_gui PUBLIC
   Qt5::Widgets
   boost_plugin_loader::boost_plugin_loader
   ${PROJECT_NAME}_analysis
   ${PROJECT_NAME}_target_finders_opencv)
 
 # add executable for widget
-add_executable(${PROJECT_NAME}_app src/ic_app.cpp)
-# target_include_directories(${PROJECT_NAME}_app PRIVATE
-#   "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>"
-#   "$<INSTALL_INTERFACE:include>")
-target_link_libraries(${PROJECT_NAME}_app PRIVATE ${PROJECT_NAME}_widget)
+add_executable(${PROJECT_NAME}_extrinsic_hand_eye_calibration_app src/app/extrinsic_hand_eye_calibration_app.cpp)
+target_link_libraries(${PROJECT_NAME}_extrinsic_hand_eye_calibration_app PRIVATE ${PROJECT_NAME}_gui)
+
+# Install headers
+install(DIRECTORY include/ DESTINATION include)
 
 # Install executables
-# install(TARGETS ${PROJECT_NAME}_app EXPORT ${PROJECT_NAME}-targets DESTINATION bin)
 configure_component(
   COMPONENT gui 
   NAMESPACE industrial_calibration 
-  TARGETS ${PROJECT_NAME}_widget ${PROJECT_NAME}_app 
+  TARGETS ${PROJECT_NAME}_gui ${PROJECT_NAME}_extrinsic_hand_eye_calibration_app
   DEPENDENCIES "Qt5 REQUIRED COMPONENTS Widgets" boost_plugin_loader)
