+++ gui/include/industrial_calibration/gui/aspect_ratio_pixmap_label.h
