+++ gui/include/industrial_calibration/gui/configurable_widget.h
