+++ gui/src/charuco_grid_target_finder.cpp
@@ -1,34 +1,36 @@
-#include "widget/aruco_grid_target_finder.h"
-#include "ui_aruco_grid_target_finder.h"
+#include <industrial_calibration/gui/charuco_grid_target_finder.h>
+#include "ui_charuco_grid_target_finder.h"
 
 namespace industrial_calibration
 {
-ArucoGridTargetFinderWidget::ArucoGridTargetFinderWidget(QWidget *parent) :
+CharucoGridTargetFinderWidget::CharucoGridTargetFinderWidget(QWidget *parent) :
     ConfigurableWidget(parent),
-    ui_(new Ui::ArucoGridTargetFinder)
+    ui_(new Ui::CharucoGridTargetFinder)
 {
     ui_->setupUi(this);
 }
 
-ArucoGridTargetFinderWidget::~ArucoGridTargetFinderWidget()
+CharucoGridTargetFinderWidget::~CharucoGridTargetFinderWidget()
 {
     delete ui_;
 }
 
-void ArucoGridTargetFinderWidget::configure(const YAML::Node& node)
+void CharucoGridTargetFinderWidget::configure(const YAML::Node& node)
 {
     ui_->rowSpinBox->setValue(node["rows"].as<int>());
     ui_->colSpinBox->setValue(node["cols"].as<int>());
+    ui_->checkerSizeDoubleSpinBox->setValue(node["chessboard_dim"].as<double>());
     ui_->markerSizeDoubleSpinBox->setValue(node["aruco_marker_dim"].as<double>());
     ui_->dictComboBox->setCurrentIndex(node["dictionary"].as<int>());
 }
 
-YAML::Node ArucoGridTargetFinderWidget::save() const
+YAML::Node CharucoGridTargetFinderWidget::save() const
 {
     YAML::Node node;
-    node["type"] = "ArucoGridTargetFinder";
+    node["type"] = "CharucoGridTargetFinder";
     node["rows"] = ui_->rowSpinBox->value();
     node["cols"] = ui_->colSpinBox->value();
+    node["chessboard_dim"] = ui_->checkerSizeDoubleSpinBox->value();
     node["aruco_marker_dim"] = ui_->markerSizeDoubleSpinBox->value();
     node["dictionary"] = ui_->dictComboBox->currentIndex();
 
@@ -36,4 +38,3 @@ YAML::Node ArucoGridTargetFinderWidget::save() const
 }
 
 } // namespace industrial_calibration
-
